PR feedback

+++ libs/simdvec/src/test/java/org/elasticsearch/simdvec/BFloat16VectorScorerFactoryTests.java
@@ -36,7 +36,7 @@ public class BFloat16VectorScorerFactoryTests extends AbstractVectorTestCase {
     private static final int TIMES = 100; // a loop iteration times
     // BFloat16 has ~7 bits of mantissa, so relative precision is ~2^-7 ≈ 0.008.
     // Accumulation over many dims can compound this; 1e-3 is a practical bound.
-    private static final double DELTA = 1e-3;
+    private static final float DELTA = 1e-3f;
 
     // Tests that the provider instance is present or not on expected platforms/architectures
     public void testSupport() {
@@ -229,7 +229,7 @@ public class BFloat16VectorScorerFactoryTests extends AbstractVectorTestCase {
 
             float[] actual = new float[size];
             scorer.bulkScore(nodes, actual, nodes.length);
-            assertArrayEqualsPercent(sim.toString(), expected, actual, 1, (float) DELTA);
+            assertArrayEqualsPercent(sim.toString(), expected, actual, 1, DELTA);
         }
     }
